requirements.txt
amqp==2.5.2
astroid==2.3.3
attrs==19.3.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.7
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.0.0
